+++ src/edu/ufp/inf/sd/rmi/froggergame/client/frogger/CollisionObject.java
@@ -23,26 +23,34 @@
  * OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package edu.ufp.inf.sd.rmi.frogger;
-
+package edu.ufp.inf.sd.rmi.froggergame.client.frogger;
+import jig.engine.physics.vpe.VanillaSphere;
 import jig.engine.util.Vector2D;
-  
-public class Truck extends MovingEntity {
-	
-	public static int LENGTH = 32*2;
 
-	public Truck (Vector2D pos, Vector2D v) {
-		super(Main.SPRITE_SHEET + "#truck");
-		position = pos;
-		Vector2D posSphere1 = position;
-		Vector2D posSphere2 = new Vector2D(position.getX()+32, position.getY());
-		collisionObjects.add(new CollisionObject(posSphere1));
-		collisionObjects.add(new CollisionObject(posSphere2));
-		velocity = v;
-		
-		if (v.getX() < 0)
-			setFrame(1);
-		else
-			setFrame(0);
+
+public class CollisionObject extends VanillaSphere {
+
+	public CollisionObject(Vector2D pos) {
+		super("col");
+		setPosition(pos);
+	}
+	
+	public CollisionObject(String name, Vector2D pos) {
+		super(name);
+		setPosition(pos);
+	}
+	
+	/**
+	 * Depending on the collision sphere, we offset it's position so
+	 * that it appears in the middle of the object
+	 */
+	public void setPosition(Vector2D pos) {
+		double dX = 16 - getRadius();
+		double dY = -getRadius() +16;
+		position = new Vector2D(pos.getX()+dX, pos.getY()+dY);
+	}
+	
+	public void update(long deltaMs) {
+		;
 	}
 }
