+++ src/edu/ufp/inf/sd/rmi/froggergame/client/frogger/MovingEntity.java
@@ -23,7 +23,7 @@
  * OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package edu.ufp.inf.sd.rmi.frogger;
+package edu.ufp.inf.sd.rmi.froggergame.client.frogger;
 
 import jig.engine.physics.Body;
 import jig.engine.util.Vector2D;
