+++ src/edu/ufp/inf/sd/rmi/froggergame/client/frogger/Truck.java
@@ -23,27 +23,26 @@
  * OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package edu.ufp.inf.sd.rmi.frogger;
+package edu.ufp.inf.sd.rmi.froggergame.client.frogger;
+
 import jig.engine.util.Vector2D;
   
-public class ShortLog extends MovingEntity {
+public class Truck extends MovingEntity {
 	
-	public static int LENGTH = 32*3;
-	
-	public ShortLog (Vector2D pos, Vector2D v) {
-		super(Main.SPRITE_SHEET + "#shortlog");
+	public static int LENGTH = 32*2;
+
+	public Truck (Vector2D pos, Vector2D v) {
+		super(Main.SPRITE_SHEET + "#truck");
 		position = pos;
 		Vector2D posSphere1 = position;
 		Vector2D posSphere2 = new Vector2D(position.getX()+32, position.getY());
-		Vector2D posSphere3 = new Vector2D(position.getX()+64, position.getY());
-		collisionObjects.add(new CollisionObject("colSmall", posSphere1));
-		collisionObjects.add(new CollisionObject("colSmall", posSphere2));
-		collisionObjects.add(new CollisionObject("colSmall", posSphere3));
+		collisionObjects.add(new CollisionObject(posSphere1));
+		collisionObjects.add(new CollisionObject(posSphere2));
 		velocity = v;
+		
 		if (v.getX() < 0)
 			setFrame(1);
 		else
 			setFrame(0);
 	}
-
 }
