+++ src/edu/ufp/inf/sd/rmi/froggergame/client/resources/frog_die.ogg
