+++ src/edu/ufp/inf/sd/rmi/froggergame/client/resources/frogger_sprites.png
