+++ src/edu/ufp/inf/sd/rmi/froggergame/client/resources/heat.ogg
